Every function in its own file, part 6 [#CLICKHOUSE-2]
