+++ dbms/src/Functions/cutFragment.cpp
@@ -1,0 +1,16 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/fragment.h>
+#include <Functions/FunctionStringToString.h>
+
+namespace DB
+{
+
+struct NameCutFragment { static constexpr auto name = "cutFragment"; };
+using FunctionCutFragment = FunctionStringToString<CutSubstringImpl<ExtractFragment<false>>, NameCutFragment>;
+
+void registerFunctionsURL(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionCutFragment>();
+}
+
+}
