+++ dbms/src/Functions/domainWithoutWWW.cpp
@@ -1,0 +1,17 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStringToString.h>
+#include <Functions/domain.h>
+
+namespace DB
+{
+
+struct NameDomainWithoutWWW { static constexpr auto name = "domainWithoutWWW"; };
+using FunctionDomainWithoutWWW = FunctionStringToString<ExtractSubstringImpl<ExtractDomain<true>>, NameDomainWithoutWWW>;
+
+
+void registerFunctionDomainWithoutWWW(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionDomainWithoutWWW>();
+}
+
+}
