+++ dbms/src/Functions/fragment.cpp
@@ -1,0 +1,16 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStringToString.h>
+#include <Functions/fragment.h>
+
+namespace DB
+{
+
+struct NameFragment { static constexpr auto name = "fragment"; };
+using FunctionFragment = FunctionStringToString<ExtractSubstringImpl<ExtractFragment<true>>, NameFragment>;
+
+void registerFunctionFragment(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionFragment>();
+}
+
+}
