+++ dbms/src/Functions/protocol.cpp
@@ -1,0 +1,17 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStringToString.h>
+#include <Functions/protocol.h>
+
+
+namespace DB
+{
+
+struct NameProtocol { static constexpr auto name = "protocol"; };
+using FunctionProtocol = FunctionStringToString<ExtractSubstringImpl<ExtractProtocol>, NameProtocol>;
+
+void registerFunctionProtocol(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionProtocol>();
+}
+
+}
