+++ dbms/src/Functions/queryString.cpp
@@ -1,0 +1,16 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStringToString.h>
+#include <Functions/queryString.h>
+
+namespace DB
+{
+
+struct NameQueryString { static constexpr auto name = "queryString"; };
+using FunctionQueryString = FunctionStringToString<ExtractSubstringImpl<ExtractQueryString<true>>, NameQueryString>;
+
+void registerFunctionQueryString(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionQueryString>();
+}
+
+}
