+++ dbms/src/Functions/queryStringAndFragment.cpp
@@ -1,0 +1,16 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStringToString.h>
+#include <Functions/queryStringAndFragment.h>
+
+namespace DB
+{
+
+struct NameQueryStringAndFragment { static constexpr auto name = "queryStringAndFragment"; };
+using FunctionQueryStringAndFragment = FunctionStringToString<ExtractSubstringImpl<ExtractQueryStringAndFragment<true>>, NameQueryStringAndFragment>;
+
+void registerFunctionQueryStringAndFragment(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionQueryStringAndFragment>();
+}
+
+}
