+++ dbms/src/Functions/registerFunctionsDateTime.cpp
@@ -1,8 +1,7 @@
-#include <Functions/FunctionFactory.h>
-
-
 namespace DB
 {
+
+class FunctionFactory;
 
 void registerFunctionToYear(FunctionFactory &);
 void registerFunctionToQuarter(FunctionFactory &);
